refactor: progress 파일 위치를 util로 변경
